make clients happy
