+++ pyproject.toml
@@ -17,6 +17,7 @@ python = "^3.7"
 base58 = "^2.1.0"
 ECPy = "^1.2.5"
 typing-extensions = "^3.7.4"
+requests = "^2.25.1"
 
 [tool.poetry.dev-dependencies]
 flake8 = "^3.8.4"
