+++ xrpl/clients/__init__.py
@@ -1,10 +1,7 @@
 from xrpl.clients.client import Client  # noqa F401
-from xrpl.clients.json_rpc_client import (  # noqa F401
-    JsonRpcClient,
-    json_to_response,
-    request_to_json_rpc,
-)
+from xrpl.clients.json_rpc_client import JsonRpcClient  # noqa F401
 from xrpl.clients.rippled_exception import RippledException  # noqa F401
+from xrpl.clients.utils import json_to_response, request_to_json_rpc
 
 __all__ = [
     "JsonRpcClient",
